13481 Unit tests for retry handling
